--- a/src/JustEat.StatsD/Buffered/Buffer.cs
+++ b/src/JustEat.StatsD/Buffered/Buffer.cs
@@ -2,15 +2,15 @@ using System;
 
 namespace JustEat.StatsD.Buffered
 {
-    internal ref struct Buffer
+    internal ref struct Buffer<T>
     {
-        public Buffer(Span<byte> source)
+        public Buffer(Span<T> source)
         {
             Tail = source;
             Written = 0;
         }
 
-        public Span<byte> Tail;
+        public Span<T> Tail;
         public int Written;
     }
 }