--- a/src/JustEat.StatsD/Buffered/BufferBasedStatsDPublisher.cs
+++ b/src/JustEat.StatsD/Buffered/BufferBasedStatsDPublisher.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace JustEat.StatsD.Buffered
 {
@@ -9,12 +10,12 @@ namespace JustEat.StatsD.Buffered
 
         [ThreadStatic]
         private static byte[]? _buffer;
-        private static byte[] Buffer => _buffer ?? (_buffer = new byte[SafeUdpPacketSize]);
+        private static byte[] Buffer => _buffer ??= new byte[SafeUdpPacketSize];
 
 #pragma warning disable CA5394
         [ThreadStatic]
         private static Random? _random;
-        private static Random Random => _random ?? (_random = new Random());
+        private static Random Random => _random ??= new Random();
 #pragma warning disable CA5394
 
         private readonly StatsDUtf8Formatter _formatter;
@@ -30,22 +31,22 @@ namespace JustEat.StatsD.Buffered
 
             _onError = configuration.OnError;
             _transport = transport;
-            _formatter = new StatsDUtf8Formatter(configuration.Prefix);
+            _formatter = new StatsDUtf8Formatter(configuration.Prefix, configuration.TagsFormatter);
         }
 
-        public void Increment(long value, double sampleRate, string bucket)
+        public void Increment(long value, double sampleRate, string bucket, Dictionary<string, string?>? tags)
         {
-            SendMessage(sampleRate, StatsDMessage.Counter(value, bucket));
+            SendMessage(sampleRate, StatsDMessage.Counter(value, bucket, tags));
         }
 
-        public void Gauge(double value, string bucket)
+        public void Gauge(double value, string bucket, Dictionary<string, string?>? tags)
         {
-            SendMessage(DefaultSampleRate, StatsDMessage.Gauge(value, bucket));
+            SendMessage(DefaultSampleRate, StatsDMessage.Gauge(value, bucket, tags));
         }
 
-        public void Timing(long duration, double sampleRate, string bucket)
+        public void Timing(long duration, double sampleRate, string bucket, Dictionary<string, string?>? tags)
         {
-            SendMessage(sampleRate, StatsDMessage.Timing(duration, bucket));
+            SendMessage(sampleRate, StatsDMessage.Timing(duration, bucket, tags));
         }
 
         private void SendMessage(double sampleRate, in StatsDMessage msg)