--- a/src/JustEat.StatsD/Buffered/StatsDMessage.cs
+++ b/src/JustEat.StatsD/Buffered/StatsDMessage.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace JustEat.StatsD.Buffered
 {
     internal readonly struct StatsDMessage
@@ -5,27 +7,42 @@ namespace JustEat.StatsD.Buffered
         public readonly string StatBucket;
         public readonly double Magnitude;
         public readonly StatsDMessageKind MessageKind;
+        public readonly Dictionary<string, string?>? Tags;
 
-        private StatsDMessage(string statBucket, double magnitude, StatsDMessageKind messageKind)
+        private StatsDMessage(
+            string statBucket,
+            double magnitude,
+            StatsDMessageKind messageKind,
+            Dictionary<string, string?>? tags)
         {
             StatBucket = statBucket;
             Magnitude = magnitude;
             MessageKind = messageKind;
+            Tags = tags;
         }
 
-        public static StatsDMessage Timing(long milliseconds, string statBucket)
+        public static StatsDMessage Timing(
+            long milliseconds,
+            string statBucket,
+            Dictionary<string, string?>? tags)
         {
-            return new StatsDMessage(statBucket, milliseconds, StatsDMessageKind.Timing);
+            return new StatsDMessage(statBucket, milliseconds, StatsDMessageKind.Timing, tags);
         }
 
-        public static StatsDMessage Counter(long magnitude, string statBucket)
+        public static StatsDMessage Counter(
+            long magnitude,
+            string statBucket,
+            Dictionary<string, string?>? tags)
         {
-            return new StatsDMessage(statBucket, magnitude, StatsDMessageKind.Counter);
+            return new StatsDMessage(statBucket, magnitude, StatsDMessageKind.Counter, tags);
         }
 
-        public static StatsDMessage Gauge(double magnitude, string statBucket)
+        public static StatsDMessage Gauge(
+            double magnitude,
+            string statBucket,
+            Dictionary<string, string?>? tags)
         {
-            return new StatsDMessage(statBucket, magnitude, StatsDMessageKind.Gauge);
+            return new StatsDMessage(statBucket, magnitude, StatsDMessageKind.Gauge, tags);
         }
     }
 }