--- a/src/JustEat.StatsD/IStatsDPublisher.cs
+++ b/src/JustEat.StatsD/IStatsDPublisher.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace JustEat.StatsD
 {
     /// <summary>
@@ -11,14 +13,16 @@ namespace JustEat.StatsD
         /// <param name="value">The value to increment the counter by.</param>
         /// <param name="sampleRate">The sample rate for the counter.</param>
         /// <param name="bucket">The bucket to increment the counter for.</param>
-        void Increment(long value, double sampleRate, string bucket);
+        /// <param name="tags">The tag(s) to publish with the counter.</param>
+        void Increment(long value, double sampleRate, string bucket, Dictionary<string, string?>? tags);
 
         /// <summary>
         /// Publishes a gauge for the specified bucket and value.
         /// </summary>
         /// <param name="value">The value to publish for the gauge.</param>
         /// <param name="bucket">The bucket to publish the gauge for.</param>
-        void Gauge(double value, string bucket);
+        /// <param name="tags">The tag(s) to publish with the gauge.</param>
+        void Gauge(double value, string bucket, Dictionary<string, string?>? tags);
 
         /// <summary>
         /// Publishes a timer for the specified bucket and value.
@@ -26,6 +30,7 @@ namespace JustEat.StatsD
         /// <param name="duration">The value to publish for the timer.</param>
         /// <param name="sampleRate">The sample rate for the timer.</param>
         /// <param name="bucket">The bucket to publish the timer for.</param>
-        void Timing(long duration, double sampleRate, string bucket);
+        /// <param name="tags">The tag(s) to publish with the timer.</param>
+        void Timing(long duration, double sampleRate, string bucket, Dictionary<string, string?>? tags);
     }
 }