--- a/src/JustEat.StatsD/StatsDConfiguration.cs
+++ b/src/JustEat.StatsD/StatsDConfiguration.cs
@@ -1,4 +1,5 @@
 using System;
+using JustEat.StatsD.TagsFormatters;
 
 namespace JustEat.StatsD
 {
@@ -50,6 +51,11 @@ namespace JustEat.StatsD
         /// Gets or sets an optional prefix to use for all stats.
         /// </summary>
         public string Prefix { get; set; } = string.Empty;
+
+        /// <summary>
+        /// Gets or sets the formatter for tags. By default an instance of <see cref="NoOpTagsFormatter"/> is used.
+        /// </summary>
+        public IStatsDTagsFormatter TagsFormatter { get; set; } = new NoOpTagsFormatter();
 
         /// <summary>
         /// Gets or sets an optional delegate to invoke when an error occurs