--- a/src/JustEat.StatsD/StatsDPublisher.cs
+++ b/src/JustEat.StatsD/StatsDPublisher.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using JustEat.StatsD.Buffered;
 using JustEat.StatsD.EndpointLookups;
 
@@ -82,21 +83,21 @@ namespace JustEat.StatsD
         }
 
         /// <inheritdoc />
-        public void Increment(long value, double sampleRate, string bucket)
+        public void Increment(long value, double sampleRate, string bucket, Dictionary<string, string?>? tags)
         {
-            _inner.Increment(value, sampleRate, bucket);
+            _inner.Increment(value, sampleRate, bucket, tags);
         }
 
         /// <inheritdoc />
-        public void Gauge(double value, string bucket)
+        public void Gauge(double value, string bucket, Dictionary<string, string?>? tags)
         {
-            _inner.Gauge(value, bucket);
+            _inner.Gauge(value, bucket, tags);
         }
 
         /// <inheritdoc />
-        public void Timing(long duration, double sampleRate, string bucket)
+        public void Timing(long duration, double sampleRate, string bucket, Dictionary<string, string?>? tags)
         {
-            _inner.Timing(duration, sampleRate, bucket);
+            _inner.Timing(duration, sampleRate, bucket, tags);
         }
 
         /// <inheritdoc />