--- a/src/JustEat.StatsD/TagsFormatters/LibratoTagsFormatter.cs
+++ b/src/JustEat.StatsD/TagsFormatters/LibratoTagsFormatter.cs
@@ -4,7 +4,7 @@ namespace JustEat.StatsD.TagsFormatters
     /// Formats StatsD tags for Librato.
     /// Tags placed right after the bucket name with format: <code>"#" + tag1=value1,tag2,tag3=value</code>.
     /// </summary>
-    public sealed class LibratoTagsFormatter : StatsDTagsFormatter
+    internal sealed class LibratoTagsFormatter : StatsDTagsFormatter
     {
         private const string Prefix = "#";
         private const bool AreTrailingTags = false;