--- a/src/JustEat.StatsD/TagsFormatters/NoOpTagsFormatter.cs
+++ b/src/JustEat.StatsD/TagsFormatters/NoOpTagsFormatter.cs
@@ -5,7 +5,7 @@ namespace JustEat.StatsD.TagsFormatters
 {
     internal sealed class NoOpTagsFormatter : IStatsDTagsFormatter
     {
-        public bool AreTrailing { get; }
+        public bool AreTrailing => false;
 
         public int GetTagsBufferSize(in Dictionary<string, string?> tags) => 0;