--- a/src/JustEat.StatsD/TagsFormatters/SignalFxTagsFormatter.cs
+++ b/src/JustEat.StatsD/TagsFormatters/SignalFxTagsFormatter.cs
@@ -4,7 +4,7 @@ namespace JustEat.StatsD.TagsFormatters
     /// Formats StatsD tags for SignalFX.
     /// Tags placed right after the bucket name with format: <code>"[" + tag1=value1,tag2,tag3=value + "]"</code>.
     /// </summary>
-    public sealed class SignalFxTagsFormatter : StatsDTagsFormatter
+    internal sealed class SignalFxTagsFormatter : StatsDTagsFormatter
     {
         private const string Prefix = "[";
         private const string Suffix = "]";