… README.md
… src/JustEat.StatsD/IStatsDPublisherExtensions.cs
… src/JustEat.StatsD/PublicAPI.Shipped.txt
… src/JustEat.StatsD/TagsFormatters/StatsDTagsFormatterConfiguration.cs
… src/JustEat.StatsD/TagsFormatters/TrailingTagsFormatter.cs
… src/JustEat.StatsD/TimerExtensions.cs
… tests/Benchmark/Benchmark.csproj
… tests/Benchmark/StatSendingBenchmark.cs
… tests/Benchmark/Utf8FormatterBenchmark.cs
… tests/JustEat.StatsD.Tests/Buffered/BufferBasedStatsDPublisherTests.cs
… tests/JustEat.StatsD.Tests/Extensions/FakeStatsPublisher.cs
… tests/JustEat.StatsD.Tests/StatsDPublisherTests.cs
… tests/JustEat.StatsD.Tests/TimerExtensionsTests.cs
… tests/JustEat.StatsD.Tests/Utf8FormatterTests.cs
… tests/JustEat.StatsD.Tests/Utf8TagsFormatterTests.cs
… version.props