Support of tags or dimensions for the following providers:

- [DataDog](https://docs.datadoghq.com/developers/dogstatsd/datagram_shell/?tab=metrics), [Splunk](https://docs.splunk.com/Documentation/Splunk/8.0.5/Metrics/GetMetricsInStatsd) and [CloudWatch](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch-Agent-custom-metrics-statsd.html).
- [InfluxDB](https://www.influxdata.com/blog/getting-started-with-sending-statsd-metrics-to-telegraf-influxdb/#introducing-influx-statsd).
- [Librato](https://github.com/librato/statsd-librato-backend#tags).
- [SignalFX](https://docs.signalfx.com/en/latest/integrations/agent/monitors/collectd-statsd.html#adding-dimensions-to-statsd-metrics).

It covers issue: #268 

It contains a breaking change. The interface `IStatsDPublisher` has been modified to include the tags as `Dictionary<string, string?>`. The old methods of this interface have been included as extension methods at `IStatsDPublisherExtensions`.
Therefore:
- Implementations of `IStatsDPublisher` will be broken after the change.
- Former usage of `IStatsDPublisher` and `StatsDPublisher` will work as it was in the previous version.

## Benchmarking

### After the change
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1016 (1909/November2018Update/19H2)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.401
  [Host]     : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT


```
|                                   Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|                              BufferBased | 248.0 ns | 4.58 ns | 4.29 ns |     - |     - |     - |         - |
|                  BufferBasedIgnoringTags | 272.0 ns | 3.61 ns | 3.01 ns |     - |     - |     - |         - |
|             BufferBasedIgnoringEmptyTags | 256.3 ns | 5.16 ns | 5.94 ns |     - |     - |     - |         - |
|  BufferBasedWithTagsFormatterAndNullTags | 248.9 ns | 4.86 ns | 5.59 ns |     - |     - |     - |         - |
|             BufferBasedWithTagsFormatter | 661.2 ns | 4.65 ns | 4.35 ns |     - |     - |     - |         - |
| BufferBasedWithTagsFormatterAndEmptyTags | 252.7 ns | 3.36 ns | 2.98 ns |     - |     - |     - |         - |


### Previous version
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1016 (1909/November2018Update/19H2)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.401
  [Host]     : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT
  DefaultJob : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT


```
|      Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|--------:|------:|------:|------:|----------:|
| BufferBased | 226.8 ns | 4.48 ns | 3.97 ns |     - |     - |     - |         - |