@jgpayvision 2021-03-06
Sorry for the delay @martincostello.

All changes uploaded except the change from `Dictionary` to `IDictionary`.