@martincostello 2021-03-08
Thanks for your contribution @gomete - we'll need to make some changes internally to prepare other libraries that depend on JustEat.StatsD due to the breaking changes to the interface, so there'll be a bit of lag on getting this actually published to NuGet.org.

I'll try and publish a 5.0.0 beta to NuGet.org at some point this week.