@jgpayvision 2021-03-08
Sure! Thank you Martin!

Kind regards,

Jorge Gomez Rivas
Software Engineer

[cid:image001.png@01D7142F.21DE8A90]

From: Martin Costello <notifications@github.com>
Sent: Monday, March 8, 2021 3:23 PM
To: justeat/JustEat.StatsD <JustEat.StatsD@noreply.github.com>
Cc: Jorge Gomez Rivas <J.Gomez@payvision.com>; Author <author@noreply.github.com>
Subject: Re: [justeat/JustEat.StatsD] Tagging (#312)


Thanks for your contribution @gomete<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fgomete&data=04%7C01%7Cj.gomez%40payvision.com%7C618b64d648f84456670d08d8e23db0e1%7C37f9569137d54403bf0fd29bb85a2533%7C0%7C0%7C637508101894563193%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=GSkO%2FOOg%2Brq5SC3jnzjV5hwRaY%2FH%2B8e8%2FTqMbItFHRU%3D&reserved=0> - we'll need to make some changes internally to prepare other libraries that depend on JustEat.StatsD due to the breaking changes to the interface, so there'll be a bit of lag on getting this actually published to NuGet.org.

I'll try and publish a 5.0.0 beta to NuGet.org at some point this week.

-
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fjusteat%2FJustEat.StatsD%2Fpull%2F312%23issuecomment-792788119&data=04%7C01%7Cj.gomez%40payvision.com%7C618b64d648f84456670d08d8e23db0e1%7C37f9569137d54403bf0fd29bb85a2533%7C0%7C0%7C637508101894573148%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=RFIpRPWpfpYoch4aHI1cRpdKSaxHhqFnzo1dt7TRzBk%3D&reserved=0>, or unsubscribe<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FASQL3UB5KK4F4BMX3TO42YDTCTMUTANCNFSM4WH26C3A&data=04%7C01%7Cj.gomez%40payvision.com%7C618b64d648f84456670d08d8e23db0e1%7C37f9569137d54403bf0fd29bb85a2533%7C0%7C0%7C637508101894573148%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=AQsKpWK0%2BqwEAgYY3doestxjNlWRVOtUPPZF2iUlC5A%3D&reserved=0>.